AUTOML/requirements.txt
pandas
h2o
autogluon.tabular[lightgbm,catboost,fastai,xgboost]==1.4.0
torch>=2.2,<3.0
typing_extensions
scikit-learn